feat: default `run_id` through default factory and remove `None` type annotation
Set the default of `run_id` using the default factory
mechanism. This allows a more declarative definition
of the model and `model_post_init` should be reserved
for setting something that requires multi-field access.
Additionally, the `None` type annotation is removed.
`run_id` was never `None` once it was set. The `None`
value was only used as an indication of absence of
user provided value and to signal the setting of a
default. Setting the default of `run_id` using the
default factory mechanism obviates the `None` type
annotation.