feat: default `file_mode` through default factory and remove its `"auto"` value option
Set the default of `file_mode` using the default
factory mechanism. This allows a more declarative
definition of the model and `model_post_init`
should be reserved for setting something that
requires multi-field access. Additionally, the
 `"auto"` value option is removed.
`file_mode` was never `"auto"` once it was set.
The `"auto"` value was only used as an indication
of absence of user provided value and to signal
the setting of a default. Setting the default of
`file_mode` using the default factory mechanism
obviates the `"auto"` value option and allows
the type annotation of the `file_mode` field
to be consistent with the possible values
that it holds in operation.