Remove mypy errors in `src/nwb2bids/testing/_assert_subdirectory_structure.py` (#241)
* fix(typing): remove in incompatible type assignment in `assert_subdirectory_structure()`

`subdirectory_path` is of type `str`
when assigned by the output of `os.walk`.
mypy complains when it is reassigned to
an object of `Path`

* style: use extra variable to avoid type incompatible reassignment

This allows lookup in the dictionary
`expected_structure` and the f-strings
used to construct error messages
to use the same object to
express the path.

* refactor: rename variable that holds `os.walk` values

The first element returned by os.walk()
is the full directory path, not just a
subdirectory name. The old name is
misleading.

* refactor: rename variables related to dir path in outputs of `os.walk()`

---------

Co-authored-by: Cody Baker <51133164+CodyCBakerPhD@users.noreply.github.com>